+++ rush/rush/typings/tsd.d.ts
@@ -2,7 +2,6 @@
 /// <reference path="builtins/builtins.d.ts" />
 /// <reference path="colors/colors.d.ts" />
 /// <reference path="es6-collections/es6-collections.d.ts" />
-/// <reference path="es6-promise/es6-promise.d.ts" />
 /// <reference path="git-repo-info/git-repo-info.d.ts" />
 /// <reference path="glob-escape/glob-escape.d.ts" />
 /// <reference path="read-package-tree/read-package-tree.d.ts" />